server/requirements_dev.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bitarray==2.7.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cytoolz==0.12.1
Django==4.1.7
djangorestframework==3.14.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
exceptiongroup==1.1.0
firebase-admin==6.1.0
frozenlist==1.3.3
gcloud==0.17.0
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
hexbytes==0.3.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
jws==0.1.3
lru-dict==1.1.8
MarkupSafe==2.1.2
msgpack==1.0.4
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
oauth2client==3.0.0
packaging==23.0
parsimonious==0.8.1
pkgutil-resolve-name==1.3.10
pluggy==1.0.0
proto-plus==1.22.2
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase==3.0.27
pyrsistent==0.19.3
pytest==7.2.1
pytest-django==4.5.2
python-dotenv==0.21.1
python-jwt==2.0.1
pytz==2022.7.1
PyYAML==6.0
requests==2.11.1
requests-toolbelt==0.7.0
rfc3986==1.5.0
rlp==2.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
toolz==0.12.0
uritemplate==4.1.1
urllib3==1.26.14
varint==1.0.2
web3==5.31.3
websockets==9.1
yarl==1.8.2
zipp==3.14.0
